refactor: Refactor the TUI code
