+++ src/aria2p/cli/commands/top.py
@@ -5,9 +5,9 @@ import sys
 from aria2p.api import API
 
 try:
-    from aria2p.interface import Interface
+    from aria2p.tui.wrapper import WrapperView
 except ImportError:
-    Interface = None  # type: ignore  # noqa: WPS440 (variable overlap)
+    WrapperView = None  # type: ignore  # noqa: WPS440 (variable overlap)
 
 
 def top(api: API) -> int:
@@ -20,13 +20,13 @@ def top(api: API) -> int:
     Returns:
         int: Always 0.
     """
-    if Interface is None:
+    if WrapperView is None:
         print(
             "The top-interface dependencies are not installed. Try running `pip install aria2p[tui]` to install them.",
             file=sys.stderr,
         )
         return 1
 
-    interface = Interface(api)
+    interface = WrapperView(api)
     success = interface.run()
     return 0 if success else 1
