+++ src/aria2p/tui/errors.py
@@ -1,0 +1,2 @@
+class Exit(Exception):
+    """A simple exception to exit the interactive interface."""
